已完成并测试功能： 1.注册时生成并存储条形码 2.根据用户id返回条形码 3.用户密码加密存储（MD5） 4.续借图书
